Source Change: Prevent moving a source to itself.
Previously:
- Failed on `rename()`.
- Proceeded to `::copy()` and then `::delete( $source )`.
- Returned `true`.

While this change returns `false` instead, doing this previously
would have deleted the destination that was being moved.

Therefore, this seems like a worthwhile change to the return value.

+++ src/wp-admin/includes/class-wp-filesystem-direct.php
@@ -367,6 +367,10 @@ class WP_Filesystem_Direct extends WP_Filesystem_Base {
 			return false;
 		}
 
+		if ( untrailingslashit( strtolower( $source ) ) === untrailingslashit( strtolower( $destination ) ) ) {
+			return false;
+		}
+
 		if ( $overwrite && $this->exists( $destination ) && ! $this->delete( $destination, true ) ) {
 			// Can't overwrite if the destination couldn't be deleted.
 			return false;
